pyproject.toml
[project]
name = "purple_swan"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "aleksey matiychenko", email = "aleksey@purpleswansoftware.com"},
]
dependencies = ["pandas>=2.3.3", "fsspec>=2025.10.0", "s3fs>=2025.10.0", "boto3>=1.40.61", "finnhub-python>=2.4.25", "pyarrow>=22.0.0", "yfinance>=0.2.66", "pyyaml>=6.0.3"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
